+++ src/components/ChangeLog.md
@@ -4,7 +4,7 @@
 
 #### Changelog:
 
-- New Component: `<RiveGradeBanded />`
+- New Component: `<RiveSimple />`
 - Updated Component: N/A
 - Bug Fix: N/A
 - Other Updated: Some updates to styles.